// StoreStatusBoard/ClientApp/src/app/models/response/status.forday.response.model.ts
import { StatusForDayModel } from "../status.forday.model";

export interface StatusForDayResponseModel {

  status7? : number;
  statusForDayModels7?: StatusForDayModel[];
  status7m?: number;
  statusForDayModels7m?: StatusForDayModel[];
  status8?: number;
  statusForDayModels8?: StatusForDayModel[];
  status8m?: number;
  statusForDayModels8m?: StatusForDayModel[];
  status9?: number;
  statusForDayModels9?: StatusForDayModel[];
  status9m?: number;
  statusForDayModels9m?: StatusForDayModel[];
  status10?: number;
  statusForDayModels10?: StatusForDayModel[];
  status10m?: number;
  statusForDayModels10m?: StatusForDayModel[];
  status11?: number;
  statusForDayModels11?: StatusForDayModel[];
  status11m?: number;
  statusForDayModels11m?: StatusForDayModel[];
  status12?: number;
  statusForDayModels12?: StatusForDayModel[];
  status12m?: number;
  statusForDayModels12m?: StatusForDayModel[];
  status13?: number;
  statusForDayModels13?: StatusForDayModel[];
  status13m?: number;
  statusForDayModels13m?: StatusForDayModel[];
  status14?: number;
  statusForDayModels14?: StatusForDayModel[];
  status14m?: number;
  statusForDayModels14m?: StatusForDayModel[];
  status15?: number;
  statusForDayModels15?: StatusForDayModel[];
  status15m?: number;
  statusForDayModels15m?: StatusForDayModel[];
  status16?: number;
  statusForDayModels16?: StatusForDayModel[];
  status16m?: number;
  statusForDayModels16m?: StatusForDayModel[];
  status17?: number;
  statusForDayModels17?: StatusForDayModel[];
  status17m?: number;
  statusForDayModels17m?: StatusForDayModel[];
  status18?: number;
  statusForDayModels18?: StatusForDayModel[];
  status18m?: number;
  statusForDayModels18m?: StatusForDayModel[];
  status19?: number;
  statusForDayModels19?: StatusForDayModel[];
  status19m?: number;
  statusForDayModels19m?: StatusForDayModel[];
  status20?: number;
  statusForDayModels20?: StatusForDayModel[];
  status20m?: number;
  statusForDayModels20m?: StatusForDayModel[];
  status21?: number;
  statusForDayModels21?: StatusForDayModel[];
  status21m?: number;
  statusForDayModels21m?: StatusForDayModel[];
  status22?: number;
  statusForDayModels22?: StatusForDayModel[];
}